+++ components/24-hours/BlinkingTick.tsx
