+++ components/NavMenu.tsx
@@ -12,31 +12,31 @@ const links = [
     links: [
       {
         name: "Premier League Prediction",
-        url: "/football",
+        url: "/pl-prediction",
       },
+      // {
+      //   name: "Random Footballer",
+      //   url: "https://randomfootballer.com",
+      // },
       {
-        name: "Random Footballer",
-        url: "https://randomfootballer.com",
-      },
-      {
-        name: "Do Not Waste Your Life",
-        url: "/dnwyl",
+        name: "Life in 24 Hours Calculator",
+        url: "/24-hours",
       },
     ],
   },
-  {
-    name: "Products",
-    links: [
-      {
-        name: "LuckyLink",
-        url: "https://luckylink.app",
-      },
-      {
-        name: "EarthWallet",
-        url: "https://earthwallet.app",
-      },
-    ],
-  },
+  // {
+  //   name: "Products",
+  //   links: [
+  //     {
+  //       name: "LuckyLink",
+  //       url: "https://luckylink.app",
+  //     },
+  //     {
+  //       name: "EarthWallet",
+  //       url: "https://earthwallet.app",
+  //     },
+  //   ],
+  // },
 ];
 
 export const NavMenu = ({ className }: { className: string }) => {
@@ -44,7 +44,7 @@ export const NavMenu = ({ className }: { className: string }) => {
     return url.startsWith("/") ? "_self" : "_blank";
   };
   return (
-    <ul className={className}>
+    <ul className={className} style={{ zIndex: 100 }}>
       {links.map((link) => (
         <li key={link.name}>
           {link.links ? (
