+++ components/pl-prediction/FootballTable.tsx
